requirements.txt
praw
openai
python-dotenv
requests
